MaterialBlending追加
OpenMPによるPhysicsの並列化追加

git-svn-id: http://springhead.info/spr2/Springhead2/trunk@7338 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Physics/PHAxisIndex.h
@@ -21,7 +21,7 @@ class AxisIndex {
 	int    idx[N];
 
 	/// idxの要素数
-	size_t sz;
+	int    sz;
 
 	/// 拘束フラグから拘束軸番号リストを作成する
 	void CreateList() {
@@ -56,7 +56,7 @@ public:
 	bool IsEnabled(int i) { return bCurr[i]; }
 
 	/// 拘束軸番号リストのサイズを取得する
-	int  size() { return (int)sz; }
+	int  size() { return sz; }
 
 	/// 拘束軸番号リストのn番目の軸番号を取得する
 	int& operator[](int n){ return idx[n]; }
